+++ app/src/routes/projects/SettingsTab/index.tsx
@@ -1,0 +1,3 @@
+export function SettingsTab() {
+    return <div className="flex h-[calc(100vh-5.5rem)] w-full">Settings</div>;
+}
