+++ app/src/routes/projects/TopBar/ModeToggle/index.tsx
@@ -1,17 +1,19 @@
-import { observer } from 'mobx-react-lite';
-import { useState } from 'react';
 import { ToggleGroup, ToggleGroupItem } from '@/components/ui/toggle-group';
 
-const ModeToggle = observer(() => {
-    const [mode, setMode] = useState<'projects' | 'settings'>('projects');
-
+const ModeToggle = ({
+    currentTab,
+    setCurrentTab,
+}: {
+    currentTab: 'projects' | 'settings';
+    setCurrentTab: (tab: 'projects' | 'settings') => void;
+}) => {
     return (
         <ToggleGroup
             type="single"
-            value={mode}
+            value={currentTab}
             onValueChange={(value) => {
                 if (value) {
-                    setMode(value as 'projects' | 'settings');
+                    setCurrentTab(value as 'projects' | 'settings');
                 }
             }}
             className="mb-3 h-12"
@@ -34,6 +36,6 @@ const ModeToggle = observer(() => {
             </ToggleGroupItem>
         </ToggleGroup>
     );
-});
+};
 
 export default ModeToggle;
